Conflicts:
	src/Engine/Flc.cpp
	src/Engine/Options.cpp
	src/Engine/Options.inc.h
	src/Engine/Screen.cpp
	src/Engine/Surface.cpp
	src/Engine/Surface.h
	src/Menu/OptionsVideoState.cpp

+++ src/Menu/OptionsVideoState.cpp
@@ -236,7 +236,7 @@ OptionsVideoState::OptionsVideoState(Game *game, OptionsOrigin origin) : Options
 	}
 #endif
 	
-	int selFilter = 0;
+	size_t selFilter = 0;
 	if (Screen::isOpenGLEnabled())
 	{
 #ifndef __NO_OPENGL
